Add back dropCollection to the Persistor.
Add documentation for count, getCollections, DropCollection, and CollectionStats.

Fixed a couple of typos.

Added tests for drop collections combined with getCollections, and also a test for CollectionStats

Fixed a few README formatting issues. Also changed write_concern to WriteConcern. But added code to make it backwards compatible if someone was already using write_concern

Needed quotes around some messages to make them show as JSON and formatted nicely

More than just quotes needed. weird way to get it be formatted and in a box is to make sure the line before has a ":"